chore: refactor and simplifying
