+++ package.json
@@ -71,6 +71,7 @@
     "@typescript-eslint/parser": "^5.53.0",
     "chalk": "4.1.2",
     "cli-table3": "0.6.3",
+    "date-fns": "^4.1.0",
     "eslint": "^8.34.0",
     "eslint-config-prettier": "^8.6.0",
     "ethers": "^5.7",
