+++ scripts/tradingLimits/utils/time.ts
@@ -1,25 +1,14 @@
+import { format, formatDistance, fromUnixTime } from 'date-fns'
+
 /**
- * Format the timestamp to absolute time using Intl.DateTimeFormat
+ * Format the timestamp to absolute time
  *
  * @param timestamp - Unix timestamp (seconds)
  * @returns Formatted date/time string
  */
 export function formatTimestamp(timestamp: number): string {
   try {
-    const date = new Date(timestamp * 1000)
-
-    // Use Intl.DateTimeFormat for consistent date/time formatting
-    const formatter = new Intl.DateTimeFormat('en-US', {
-      month: '2-digit',
-      day: '2-digit',
-      year: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit',
-      second: '2-digit',
-      hour12: true,
-    })
-
-    return formatter.format(date)
+    return format(fromUnixTime(timestamp), 'MM/dd/yyyy, hh:mm:ss a')
   } catch (error) {
     console.error(`Error formatting timestamp: ${timestamp}`)
     return 'Invalid date'
@@ -28,7 +17,6 @@ export function formatTimestamp(timestamp: number): string {
 
 /**
  * Format a time difference in seconds to a human-readable relative format
- * using the Intl.RelativeTimeFormat API
  *
  * @param seconds - Seconds from now to format
  * @returns Human-readable relative time
@@ -38,34 +26,8 @@ export function formatRelativeTime(seconds: number): string {
     return 'Now'
   }
 
-  // Create formatter with numeric style for brevity
-  const formatter = new Intl.RelativeTimeFormat('en', {
-    style: 'long',
-    numeric: 'always',
-  })
-
-  // Convert seconds to the appropriate unit
-  const units: [string, number][] = [
-    ['year', 60 * 60 * 24 * 365],
-    ['month', 60 * 60 * 24 * 30],
-    ['week', 60 * 60 * 24 * 7],
-    ['day', 60 * 60 * 24],
-    ['hour', 60 * 60],
-    ['minute', 60],
-    ['second', 1],
-  ]
-
-  // Find the first appropriate unit
-  for (const [unit, secondsInUnit] of units) {
-    if (seconds >= secondsInUnit || unit === 'second') {
-      const value = Math.round(seconds / secondsInUnit)
-
-      return formatter.format(value, unit as Intl.RelativeTimeFormatUnit)
-    }
-  }
-
-  // Fallback (shouldn't reach here)
-  return `in ${seconds}s`
+  const futureDate = new Date(Date.now() + seconds * 1000)
+  return `in ${formatDistance(futureDate, new Date(), { addSuffix: false })}`
 }
 
 /**
